Added -i switch for base64 decode
Temporarily commenting out the docker portion of script to test ssh key install on build agent

Troubleshooting code

Checking in scripts that are used in demo

Removing the bash script section that checks to see if hello-world container is running after running the container.  This code really doesn't apply to hello-world

Docker Image variable was not named correctly in deploy script.  Added checking to see if the container is running after the container is started back to script

Using the samples:apsnetcore app as the container for demo

Docker run --name switch was mistakenly removed from bash script

Final check-in of bash scripts used to install an ssh private key on an Azure DevOps Hosted Agent and to deploy a container to an Azure Virtual Machine.